Addition of scaffolding for Multi-functional search
This is an incomplete update, insofar as the app isn't working as it eventually will work. However, the framework is in place for implementing a drop down menu where a user can choose to search by Site Number, Place Name, or Coordiantes. Still a lot of cleanup and work to do.